python_requirements_lock.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.0
databricks-cli==0.17.8
docker==6.1.3
entrypoints==0.4
Flask==2.3.3
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mlflow==2.7.1
numpy==1.26.0
oauthlib==3.2.2
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pip==23.2.1
protobuf==4.24.3
pyarrow==13.0.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
setuptools==65.5.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.21
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.3
Werkzeug==3.0.0
zipp==3.17.0
